Force int value on model's count

+++ classes/Model.php
@@ -23,7 +23,7 @@ abstract class Model {
     }
 
     public static function count($where_sql = false, $params = []) {
-      return SQL::value('SELECT COUNT(1) FROM ' . static::persistenceOptions('table') . ($where_sql ? " where {$where_sql}" : ''), $params);
+      return (int) SQL::value('SELECT COUNT(1) FROM ' . static::persistenceOptions('table') . ($where_sql ? " where {$where_sql}" : ''), $params);
     }
 
     public static function all($page=1, $limit=-1){
